% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_reg-fit.R
\name{autoplot.brulee_linear_reg}
\alias{autoplot.brulee_linear_reg}
\title{Plot model loss over epochs}
\usage{
\method{autoplot}{brulee_linear_reg}(object, ...)
}
\arguments{
\item{object}{A \code{brulee_linear_reg} object.}

\item{...}{Not currently used}
}
\value{
A \code{ggplot} object.
}
\description{
Plot model loss over epochs
}
\details{
This function plots the loss function across the available epochs.
}
