% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lantern_logistic_reg-fit.R
\name{autoplot.lantern_logistic_reg}
\alias{autoplot.lantern_logistic_reg}
\title{Plot model loss over epochs}
\usage{
\method{autoplot}{lantern_logistic_reg}(object, ...)
}
\arguments{
\item{object}{A `lantern_logistic_reg` object.}

\item{...}{Not currently used}
}
\value{
A `ggplot` object.
}
\description{
Plot model loss over epochs
}
\details{
This function plots the loss function across the available epochs.
}
