% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lantern_mlp-fit.R
\name{autoplot.lantern_mlp}
\alias{autoplot.lantern_mlp}
\title{Plot model loss over epochs}
\usage{
\method{autoplot}{lantern_mlp}(object, ...)
}
\arguments{
\item{object}{A \code{lantern_mlp} object.}

\item{...}{Not currently used}
}
\value{
A \code{ggplot} object.
}
\description{
Plot model loss over epochs
}
\details{
This function plots the loss function across the available epochs.
}
