% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{brulee-autoplot}
\alias{brulee-autoplot}
\alias{autoplot.brulee_mlp}
\alias{autoplot.brulee_logistic_reg}
\alias{autoplot.brulee_multinomial_reg}
\alias{autoplot.brulee_linear_reg}
\title{Plot model loss over epochs}
\usage{
\method{autoplot}{brulee_mlp}(object, ...)

\method{autoplot}{brulee_logistic_reg}(object, ...)

\method{autoplot}{brulee_multinomial_reg}(object, ...)

\method{autoplot}{brulee_linear_reg}(object, ...)
}
\arguments{
\item{object}{A \code{brulee_mlp}, \code{brulee_logistic_reg},
\code{brulee_multinomial_reg}, or \code{brulee_linear_reg} object.}

\item{...}{Not currently used}
}
\value{
A \code{ggplot} object.
}
\description{
Plot model loss over epochs
}
\details{
This function plots the loss function across the available epochs.
}
