% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{brulee-coefs}
\alias{brulee-coefs}
\alias{coef.brulee_logistic_reg}
\alias{coef.brulee_linear_reg}
\alias{coef.brulee_mlp}
\alias{coef.brulee_multinomial_reg}
\title{Extract Model Coefficients}
\usage{
\method{coef}{brulee_logistic_reg}(object, epoch = NULL, ...)

\method{coef}{brulee_linear_reg}(object, epoch = NULL, ...)

\method{coef}{brulee_mlp}(object, epoch = NULL, ...)

\method{coef}{brulee_multinomial_reg}(object, epoch = NULL, ...)
}
\arguments{
\item{object}{A model fit from \pkg{brulee}.}

\item{epoch}{A single integer for the training iteration. If left \code{NULL},
the estimates from the best model fit (via internal performance metrics).}

\item{...}{Not currently used.}
}
\value{
For logistic/linear regression, a named vector. For neural networks,
a list of arrays.
}
\description{
Extract Model Coefficients
}
