% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{brulee-internal}
\alias{brulee-internal}
\alias{tunable.brulee_mlp}
\alias{tunable.brulee_logistic_reg}
\alias{tunable.brulee_multinomial_reg}
\alias{tunable.brulee_linear_reg}
\title{Internal functions and methods}
\usage{
\method{tunable}{brulee_mlp}(x, ...)

\method{tunable}{brulee_logistic_reg}(x, ...)

\method{tunable}{brulee_multinomial_reg}(x, ...)

\method{tunable}{brulee_linear_reg}(x, ...)
}
\description{
Internal functions and methods
}
\keyword{internal}
