% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brulee-package.R
\docType{package}
\name{brulee-package}
\alias{brulee}
\alias{brulee-package}
\title{brulee: High-Level Modeling Functions with 'torch'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides high-level modeling functions to define and train models using the 'torch' R package. Models include linear, logistic, and multinomial regression, multilayer perceptrons, modern neural neural networks, and the pretrained tabular foundational models such as 'TabICL' and 'TabPFN'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/brulee}
  \item \url{https://brulee.tidymodels.org/}
  \item Report bugs at \url{https://github.com/tidymodels/brulee/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
  \item Daniel Falbel \email{daniel@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Prior Labs GmbH (TabPFN code ported from the Python 'tabpfn' package; see inst/COPYRIGHTS) [copyright holder]
}

}
\keyword{internal}
