% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resnet-fit.R
\name{brulee_resnet}
\alias{brulee_resnet}
\alias{brulee_resnet.default}
\alias{brulee_resnet.data.frame}
\alias{brulee_resnet.matrix}
\alias{brulee_resnet.formula}
\alias{brulee_resnet.recipe}
\title{Fit residual neural networks (ResNet)}
\usage{
brulee_resnet(x, ...)

\method{brulee_resnet}{default}(x, ...)

\method{brulee_resnet}{data.frame}(
  x,
  y,
  epochs = 100L,
  hidden_units = 3L,
  num_layers = 2L,
  block_units = 10L,
  activation = "relu",
  penalty = 0.001,
  mixture = 0,
  dropout = 0,
  validation = 0.1,
  optimizer = "LBFGS",
  learn_rate = 0.01,
  rate_schedule = "none",
  momentum = 0,
  batch_size = NULL,
  class_weights = NULL,
  stop_iter = 5,
  grad_value_clip = 5,
  grad_norm_clip = 5,
  verbose = FALSE,
  ...
)

\method{brulee_resnet}{matrix}(
  x,
  y,
  epochs = 100L,
  hidden_units = 3L,
  num_layers = 2L,
  block_units = 10L,
  activation = "relu",
  penalty = 0.001,
  mixture = 0,
  dropout = 0,
  validation = 0.1,
  optimizer = "LBFGS",
  learn_rate = 0.01,
  rate_schedule = "none",
  momentum = 0,
  batch_size = NULL,
  class_weights = NULL,
  stop_iter = 5,
  grad_value_clip = 5,
  grad_norm_clip = 5,
  verbose = FALSE,
  ...
)

\method{brulee_resnet}{formula}(
  formula,
  data,
  epochs = 100L,
  hidden_units = 3L,
  num_layers = 2L,
  block_units = 10L,
  activation = "relu",
  penalty = 0.001,
  mixture = 0,
  dropout = 0,
  validation = 0.1,
  optimizer = "LBFGS",
  learn_rate = 0.01,
  rate_schedule = "none",
  momentum = 0,
  batch_size = NULL,
  class_weights = NULL,
  stop_iter = 5,
  grad_value_clip = 5,
  grad_norm_clip = 5,
  verbose = FALSE,
  ...
)

\method{brulee_resnet}{recipe}(
  x,
  data,
  epochs = 100L,
  hidden_units = 3L,
  num_layers = 2L,
  block_units = 10L,
  activation = "relu",
  penalty = 0.001,
  mixture = 0,
  dropout = 0,
  validation = 0.1,
  optimizer = "LBFGS",
  learn_rate = 0.01,
  rate_schedule = "none",
  momentum = 0,
  batch_size = NULL,
  class_weights = NULL,
  stop_iter = 5,
  grad_value_clip = 5,
  grad_norm_clip = 5,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}

The predictor data should be standardized (e.g. centered or scaled).}

\item{...}{Options to pass to the learning rate schedulers via
\code{\link[=set_learn_rate]{set_learn_rate()}}. For example, the \code{reduction} or \code{steps} arguments to
\code{\link[=schedule_step]{schedule_step()}} could be passed here.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 column (numeric or factor).
\item A \strong{matrix} with numeric column  (numeric or factor).
\item A  \strong{vector}  (numeric or factor).
}}

\item{epochs}{An integer for the number of epochs of training.}

\item{hidden_units}{An integer for the number of hidden units, or a vector
of integers. If a vector of integers, the model will have
\code{length(hidden_units)} layers each with \code{hidden_units[i]} hidden units. If
a single integer is passed and \code{num_layers > 1}, the value is used for all
layers.}

\item{num_layers}{An integer for the number of layers within each residual block.
Must be >= 1.}

\item{block_units}{An integer for the number of hidden units in each layer
within a residual block. Must be >= 2.}

\item{activation}{A character vector for the activation function (such as
"relu", "tanh", "sigmoid", and so on). See \code{\link[=brulee_activations]{brulee_activations()}} for
a list of possible values. If \code{hidden_units} is a vector, \code{activation}
can be a character vector with length equals to \code{length(hidden_units)}
specifying the activation for each hidden layer.}

\item{penalty}{The amount of weight decay (i.e., L2 regularization).}

\item{mixture}{Proportion of Lasso Penalty (type: double, default: 0.0). A
value of mixture = 1 corresponds to a pure lasso model, while mixture = 0
indicates ridge regression (a.k.a weight decay). Must be zero for
optimizers \code{"ADAMw"}, \code{"RMSprop"}, \code{"Adadelta"}.}

\item{dropout}{The proportion of parameters set to zero.}

\item{validation}{The proportion of the data randomly assigned to a
validation set.}

\item{optimizer}{The method used in the optimization procedure. Possible choices
are \code{"SGD"},  \code{"ADAMw"}, \code{"Adadelta"}, \code{"Adagrad"}, \code{"RMSprop"}, and
\code{"LBFGS"}. \code{"LBFGS"} is the only second-order method and does not use
batches.}

\item{learn_rate}{A positive number that controls the initial rapidity that
the model moves along the descent path. Values around 0.1 or less are
typical.}

\item{rate_schedule}{A single character value for how the learning rate
should change as the optimization proceeds. Possible values are
\code{"none"} (the default), \code{"decay_time"}, \code{"decay_expo"}, \code{"cyclic"} and
\code{"step"}. See \code{\link[=schedule_decay_time]{schedule_decay_time()}} for more details.}

\item{momentum}{A positive number usually on \verb{[0.50, 0.99]} for the momentum
parameter in gradient descent.  (optimizers \code{"SGD"},  and \code{"RMSprop"} only,
ignored otherwise).}

\item{batch_size}{An integer for the number of training set points in each
batch. (\code{optimizer != "LBFGS"} only, ignored otherwise)}

\item{class_weights}{Numeric class weights (classification only). The value
can be:
\itemize{
\item A named numeric vector (in any order) where the names are the outcome
factor levels.
\item An unnamed numeric vector assumed to be in the same order as the outcome
factor levels.
\item A single numeric value for the least frequent class in the training data
and all other classes receive a weight of one.
}}

\item{stop_iter}{A non-negative integer for how many iterations with no
improvement before stopping.}

\item{grad_norm_clip, grad_value_clip}{Two numeric values, possibly \code{Inf},
that prevents the gradient's values or norm(s) from exceeding the specified
value. This can be helpful if training stops early with the message that
\code{"Loss is NaN at epoch x Training is stopped."}}

\item{verbose}{A logical that prints out the iteration history.}

\item{formula}{A formula specifying the outcome term(s) on the left-hand side,
and the predictor term(s) on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{brulee_resnet} object with elements:
\itemize{
\item \code{models_obj}: a serialized raw vector for the torch module.
\item \code{estimates}: a list of matrices with the model parameter estimates per
epoch.
\item \code{best_epoch}: an integer for the epoch with the smallest loss.
\item \code{loss}: A vector of loss values (MSE for regression, negative log-
likelihood for classification) at each epoch.
\item \code{dim}: A list of data dimensions.
\item \code{y_stats}: A list of summary statistics for numeric outcomes.
\item \code{parameters}: A list of some tuning parameter values.
\item \code{blueprint}: The \code{hardhat} blueprint data.
}
}
\description{
\code{brulee_resnet()} fits residual network models with skip connections.
}
\details{
This function fits residual network (ResNet) models for regression (when
the outcome is a number) or classification (a factor). ResNets use skip
connections that add the input of a block to its output, allowing gradients
to flow more easily through deep networks. For regression, the mean squared
error is optimized and cross-entropy is the loss function for classification.

The network consists of \code{hidden_units} residual blocks. Each block contains
\code{num_layers} layers, with each layer having \code{block_units} hidden units.
Skip connections add the block input to the block output.

This model is similar to the approach described in Equation 2 of Gorishniy
\emph{et al}. (2021). It uses a series of blocks, where batch normalization is
applied first. These values are connected to a layer of hidden units via a
nonlinear activation (ReLU by default), and new representation values are
created from them and a skip connection from the normalized values.

In this implementation, batch normalization does not require the same number
of outputs as inputs (though that is the default). Fewer outputs, called
block units here, can result in a "bottleneck" architecture described in He
\emph{et al}. (2016) if the units are set to \code{narrow -> wide -> narrow}.
Alternatively, an approach similar to Sandler \emph{et al}. (2018) called the
"inverted residual block (\code{wide -> narrow -> wide}) can be specified.

When the outcome is a number, the function internally standardizes the
outcome data to have mean zero and a standard deviation of one. The prediction
function creates predictions on the original scale.

By default, training halts when the validation loss increases for at least
\code{step_iter} iterations. If \code{validation = 0} the training set loss is used.

The \emph{predictors} data should all be numeric and encoded in the same units (e.g.
standardized to the same range or distribution). If there are factor
predictors, use a recipe or formula to create indicator variables (or some
other method) to make them numeric. Predictors should be in the same units
before training.

The model objects are saved for each epoch so that the number of epochs can
be efficiently tuned. Both the \code{\link[=coef]{coef()}} and \code{\link[=predict]{predict()}} methods for this
model have an \code{epoch} argument (which defaults to the epoch with the best
loss value).

The use of the L1 penalty (a.k.a. the lasso penalty) does \emph{not} force
parameters to be strictly zero (as it does in packages such as \pkg{glmnet}).
The zeroing out of parameters is a specific feature the optimization method
used in those packages.
\subsection{Learning Rates}{

The learning rate can be set to constant (the default) or dynamically set
via a learning rate scheduler (via the \code{rate_schedule}). Using
\code{rate_schedule = 'none'} uses the \code{learn_rate} argument. Otherwise, any
arguments to the schedulers can be passed via \code{...}.
}
}
\examples{
\dontshow{if (!brulee:::is_cran_check()) withAutoprint(\{ # examplesIf}
\donttest{
if (torch::torch_is_installed() & rlang::is_installed(c("recipes", "yardstick", "modeldata"))) {

 ## -----------------------------------------------------------------------------
 # regression examples (increase # epochs to get better results)

 data(ames, package = "modeldata")

 ames$Sale_Price <- log10(ames$Sale_Price)

 set.seed(122)
 in_train <- sample(1:nrow(ames), 2000)
 ames_train <- ames[ in_train,]
 ames_test  <- ames[-in_train,]

 # Using recipe
 library(recipes)

 ames_rec <-
  recipe(Sale_Price ~ Longitude + Latitude, data = ames_train) |>
    step_normalize(all_numeric_predictors())

 set.seed(2)
 fit <- brulee_resnet(ames_rec, data = ames_train,
                      hidden_units = 3, num_layers = 2, block_units = 10,
                      epochs = 50, batch_size = 32)
 fit

 autoplot(fit)

 library(yardstick)
 predict(fit, ames_test) |>
   bind_cols(ames_test) |>
   rmse(Sale_Price, .pred)

 # ------------------------------------------------------------------------------
 # classification

 library(dplyr)

 data("parabolic", package = "modeldata")

 set.seed(1)
 in_train <- sample(1:nrow(parabolic), 300)
 parabolic_tr <- parabolic[ in_train,]
 parabolic_te <- parabolic[-in_train,]

 set.seed(2)
 cls_fit <- brulee_resnet(class ~ ., data = parabolic_tr,
                          hidden_units = 2, num_layers = 2, block_units = 5,
                          epochs = 200L, learn_rate = 0.1, activation = "elu",
                          penalty = 0.1, batch_size = 2^8, optimizer = "SGD")
 autoplot(cls_fit)

 predict(cls_fit, parabolic_te, type = "prob") |>
   bind_cols(parabolic_te) |>
   roc_auc(class, .pred_Class1)

 }
}
\dontshow{\}) # examplesIf}
}
\references{
He, K., Zhang, X., Ren, S., & Sun, J. (2016). Deep residual learning for
image recognition. In \emph{Proceedings of the IEEE conference on computer vision
and pattern recognition} (pp. 770-778).

He, K., Zhang, X., Ren, S., & Sun, J. (2016). Identity mappings in deep
residual networks. In \emph{European conference on computer vision} (pp. 630-645).
Springer, Cham.

Gorishniy, Y., Rubachev, I., Khrulkov, V., & Babenko, A. (2021).
Revisiting deep learning models for tabular data. \emph{Advances in neural
information processing systems}, 34, 18932-18943.

Sandler, M., Howard, A., Zhu, M., Zhmoginov, A., & Chen, L. C. (2018).
Mobilenetv2: Inverted residuals and linear bottlenecks. In \emph{Proceedings of
the IEEE conference on computer vision and pattern recognition} (pp.
4510-4520).
}
\seealso{
\code{\link[=predict.brulee_resnet]{predict.brulee_resnet()}}, \code{\link[=coef.brulee_resnet]{coef.brulee_resnet()}},
\code{\link[=autoplot.brulee_resnet]{autoplot.brulee_resnet()}}
}
