% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{guess_brulee_device}
\alias{guess_brulee_device}
\title{Determine an appropriate computational device for torch}
\usage{
guess_brulee_device()
}
\value{
A character string, one of: \code{"cpu"}, \code{"cuda"}, or \code{"mps"}.
}
\description{
Uses \pkg{torch} functions to determine if there is a GPU available for use.
}
\examples{
guess_brulee_device()
}
