% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{lantern-coefs}
\alias{lantern-coefs}
\alias{coef.lantern_logistic_reg}
\alias{coef.lantern_linear_reg}
\alias{coef.lantern_mlp}
\title{Extract Model Coefficients}
\usage{
\method{coef}{lantern_logistic_reg}(object, epoch = NULL, ...)

\method{coef}{lantern_linear_reg}(object, epoch = NULL, ...)

\method{coef}{lantern_mlp}(object, epoch = NULL, ...)
}
\arguments{
\item{object}{A model fit from \pkg{lantern}.}

\item{epoch}{A single integer for the training iteration. If left \code{NULL},
the estimates from the best model fit (via internal performance metrics).}
}
\value{
For logistic/linear regression, a named vector. For neural networks,
a list of arrays.
}
\description{
Extract Model Coefficients
}
