% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{lantern-internal}
\alias{lantern-internal}
\alias{tunable.lantern_mlp}
\alias{tunable.lantern_logistic_reg}
\alias{tunable.lantern_linear_reg}
\title{Internal functions and methods}
\usage{
\method{tunable}{lantern_mlp}(x, ...)

\method{tunable}{lantern_logistic_reg}(x, ...)

\method{tunable}{lantern_linear_reg}(x, ...)
}
\description{
Internal functions and methods
}
\keyword{internal}
