% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lantern-package.R
\docType{package}
\name{lantern-package}
\alias{lantern}
\alias{lantern-package}
\title{lantern: High-Level Modeling Functions with 'torch'}
\description{
High-level modeling functions with 'torch'.
}
\author{
\strong{Maintainer}: Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
