% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg-predict.R
\name{predict.brulee_logistic_reg}
\alias{predict.brulee_logistic_reg}
\title{Predict from a \code{brulee_logistic_reg}}
\usage{
\method{predict}{brulee_logistic_reg}(object, new_data, type = NULL, epoch = NULL, ...)
}
\arguments{
\item{object}{A \code{brulee_logistic_reg} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"class"} for hard class predictions
\item \code{"prob"} for soft class predictions (i.e., class probabilities)
}}

\item{epoch}{An integer for the epoch to make predictions from. If this value
is larger than the maximum number that was fit, a warning is issued and the
parameters from the last epoch are used.}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{brulee_logistic_reg}
}
