% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rln-predict.R
\name{predict.brulee_rln}
\alias{predict.brulee_rln}
\title{Predict from a \code{brulee_rln}}
\usage{
\method{predict}{brulee_rln}(object, new_data, type = NULL, epoch = NULL, ...)
}
\arguments{
\item{object}{A \code{brulee_rln} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The only valid option is \code{"numeric"} for
numeric predictions.}

\item{epoch}{An integer for the epoch to make predictions. If larger than
the number of epochs fit, a warning is issued and the last epoch is used.
If \code{NULL} (default), the epoch with the smallest loss is used.}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions with the same number of rows as \code{new_data}.
}
\description{
Predict from a \code{brulee_rln}
}
\examples{
\dontshow{if (!brulee:::is_cran_check()) withAutoprint(\{ # examplesIf}
\donttest{
if (torch::torch_is_installed() & rlang::is_installed(c("recipes", "modeldata"))) {

 data(ames, package = "modeldata")
 ames$Sale_Price <- log10(ames$Sale_Price)

 set.seed(1)
 in_train <- sample(1:nrow(ames), 2000)
 ames_train <- ames[ in_train,]
 ames_test  <- ames[-in_train,]

 library(recipes)

 ames_rec <-
  recipe(Sale_Price ~ Longitude + Latitude, data = ames_train) |>
    step_normalize(all_numeric_predictors())

 set.seed(2)
 fit <- brulee_rln(ames_rec, data = ames_train, hidden_units = 20L, epochs = 30L)

 predict(fit, ames_test)
}
}
\dontshow{\}) # examplesIf}
}
