% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saint-predict.R
\name{predict.brulee_saint}
\alias{predict.brulee_saint}
\title{Predict from a \code{brulee_saint}}
\usage{
\method{predict}{brulee_saint}(object, new_data, type = NULL, epoch = NULL, ...)
}
\arguments{
\item{object}{A \code{brulee_saint} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
\item \code{"class"} for hard class predictions
\item \code{"prob"} for soft class predictions (i.e., class probabilities)
}}

\item{epoch}{An integer for the epoch to make predictions. If this value
is larger than the maximum number that was fit, a warning is issued and the
parameters from the last epoch are used. If left \code{NULL}, the epoch
associated with the smallest loss is used.}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{brulee_saint}
}
\examples{
\dontshow{if (!brulee:::is_cran_check()) withAutoprint(\{ # examplesIf}
\donttest{
if (torch::torch_is_installed() & rlang::is_installed(c("recipes", "modeldata"))) {
  set.seed(87261)
  tr_data <- modeldata::sim_classification(500)
  te_data <- modeldata::sim_classification(50)

  set.seed(2)
  fit <- brulee_saint(class ~ ., data = tr_data,
                      epochs = 50L, batch_size = 64L, stop_iter = 10L,
                      learn_rate = 0.001)
  fit

  autoplot(fit)

 predict(fit, te_data)
 predict(fit, te_data, type = "prob")
}
}
\dontshow{\}) # examplesIf}
}
