% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lantern_logistic_reg-predict.R
\name{predict.lantern_logistic_reg}
\alias{predict.lantern_logistic_reg}
\title{Predict from a `lantern_logistic_reg`}
\usage{
\method{predict}{lantern_logistic_reg}(object, new_data, type = NULL, epoch = NULL, ...)
}
\arguments{
\item{object}{A `lantern_logistic_reg` object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:

- `"class"` for hard class predictions
- `"prob"` for soft class predictions (i.e., class probabilities)}

\item{epoch}{An integer for the epoch to make predictions from. If this value
is larger than the maximum number that was fit, a warning is issued and the
parameters from the last epoch are used.}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in `new_data`.
}
\description{
Predict from a `lantern_logistic_reg`
}
