% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_mlp-predict.R
\name{predict.torch_mlp}
\alias{predict.torch_mlp}
\title{Predict from a \code{torch_mlp}}
\usage{
\method{predict}{torch_mlp}(object, new_data, type = "numeric", epoch = NULL, ...)
}
\arguments{
\item{object}{A \code{torch_mlp} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{epoch}{An integer for the epoch to make predictions from. If this value
is larger than the maximum number that was fit, a warning is issued and the
parameters from the last epoch are used.}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{torch_mlp}
}
