% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{tunable}
\alias{augment}
\alias{coef}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:augment]{augment()}}, \code{\link[generics:tunable]{tunable()}}}

  \item{ggplot2}{\code{\link[ggplot2:autoplot]{autoplot()}}}

  \item{stats}{\code{\link[stats:coef]{coef()}}}
}}

