% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_mlp-fit.R
\name{torch_mlp}
\alias{torch_mlp}
\alias{torch_mlp.default}
\alias{torch_mlp.data.frame}
\alias{torch_mlp.matrix}
\alias{torch_mlp.formula}
\alias{torch_mlp.recipe}
\title{Fit a single layer neural network using torch}
\usage{
torch_mlp(x, ...)

\method{torch_mlp}{default}(x, ...)

\method{torch_mlp}{data.frame}(
  x,
  y,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learning_rate = 0.01,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

\method{torch_mlp}{matrix}(
  x,
  y,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learning_rate = 0.01,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

\method{torch_mlp}{formula}(
  formula,
  data,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learning_rate = 0.01,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

\method{torch_mlp}{recipe}(
  x,
  data,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learning_rate = 0.01,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}

The predictor data should be standardized (e.g. centered or scaled).}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{epochs}{An integer for the number of epochs of training.}

\item{hidden_units}{An integer for the number of hidden units.}

\item{learning_rate}{A positive number (usually less than 0.1).}

\item{conv_crit}{A non-negative number for convergence.}

\item{verbose}{A logical that prints out the iteration history.}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{torch_mlp} object.
}
\description{
\code{torch_mlp()} fits a model.
}
