% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoint-predict.R
\name{predict.brulee_auto_int}
\alias{predict.brulee_auto_int}
\title{Predict from a \code{brulee_auto_int}}
\usage{
\method{predict}{brulee_auto_int}(object, new_data, type = NULL, epoch = NULL, ...)
}
\arguments{
\item{object}{A \code{brulee_auto_int} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
\item \code{"class"} for hard class predictions
\item \code{"prob"} for soft class predictions (i.e., class probabilities)
}}

\item{epoch}{An integer for the epoch to make predictions. If this value
is larger than the maximum number that was fit, a warning is issued and the
parameters from the last epoch are used. If left \code{NULL}, the epoch
associated with the smallest loss is used.}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{brulee_auto_int}
}
\examples{
\dontshow{if (!brulee:::is_cran_check()) withAutoprint(\{ # examplesIf}
\donttest{
if (torch::torch_is_installed() & rlang::is_installed(c("recipes", "modeldata"))) {
 data(credit_data, package = "modeldata")
 credit_data <- na.omit(credit_data)

 set.seed(1)
 in_train <- sample(1:nrow(credit_data), 3000)
 credit_train <- credit_data[ in_train,]
 credit_test  <- credit_data[-in_train,]

 set.seed(2)
 fit <- brulee_auto_int(Status ~ ., data = credit_train,
                       epochs = 50, batch_size = 256)

 predict(fit, credit_test)
 predict(fit, credit_test, type = "prob")
}
}
\dontshow{\}) # examplesIf}
}
