% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R, R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tunable}
\alias{\%>\%}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{generics}{\code{\link[generics]{tunable}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

